refactor(bds): rename components/ui/ServiceBadge/ → components/ui/ServiceTag/ (closes #731)
Post-#572 the dir hosted only ServiceTag.{tsx,css,mdx,stories.tsx} +
service-config.ts + icons/ but kept the legacy dir name as a follow-up.
Rename completes the cleanup so the dir name matches the canonical
component.

### Renames

git mv components/ui/ServiceBadge → components/ui/ServiceTag (single rename,
~50 file renames including icon SVGs preserved as renames via git's
similarity detection).

### Path updates

- `components/ui/index.ts` — `export * from './ServiceBadge'` → `'./ServiceTag'`
- `components/ui/PageHeader/PageHeader.stories.tsx` — `import { ServiceTag } from '../ServiceBadge'` → `'../ServiceTag'`
- `content-system/blueprints/react/HeroSplitImageCardOverlay.tsx` — deep path update
- `components/ui/ServiceTag/service-config.ts` — docstring updated
- `components/ui/ServiceTag/ServiceTag.mdx` — internal prose ref to `components/ui/ServiceTag/icons/...`
- `docs-site/content/docs/primitives/naming-conventions.mdx` — GitHub source link

### Workaround removed

`scripts/lint-jsdoc.js` MULTI_EXPORT no longer needs the
`ServiceBadge: ['ServiceTag']` entry (added in PR #730 as a temporary
mapping). Lint now finds `ServiceTag.tsx` via the default `${dirName}.tsx`
pattern.

### Pre-existing MDX violations fixed (revealed by rename)

The recipe lint scans `components/ui/*/${dirName}.mdx`. Pre-rename, the
dir was `ServiceBadge/` and the file `ServiceTag.mdx` — name mismatch
meant the file was invisible to the lint. Post-rename it's now scanned
and surfaced 7 pre-existing violations:

- 5× `---` dividers → removed
- Missing `<ComponentLinks slug="service-tag" />` → added
- `## Usage` H2 section (banned per ADR-007) → removed (Canvas blocks
  already show source via Storybook's "Show code" toggle)
- Title casing `# Service Tag` → `# Service tag` (sentence case matches
  Banner, TextInput conventions)

### Verification

- `npm run typecheck` clean
- `npm run lint-storybook-recipe --enforce` → 82/82 conforming (was 81;
  ServiceTag.mdx is now visible to the lint)
- `npm run lint-jsdoc` clean
- `npm run build-storybook` builds cleanly
- `node scripts/build-inspector-manifest.mjs` emits `bds-service-tag`
  → `components-service-tag--overview` (resolves in Chromatic)
- 93/94 manifest URLs resolve (only AnimatedIcon broken, tracked in
  #726 / PR #732)
