+++ docs-site/content/docs/primitives/naming-conventions.mdx
@@ -351,7 +351,7 @@ The **machine-readable source of truth** for every component's `variant`, `appea
 | IconButton | `primary` · `outline` · `secondary` · `ghost` (default) · `inverse` · `danger` · `danger-outline` · `danger-ghost` · `destructive` · `positive` · `selected` | — | shares `ButtonSize` | — | [`IconButton.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/Button/IconButton.tsx) |
 | Chip | `primary` · `secondary` | — | `sm` · `md` · `lg` | — | [`Chip.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/Chip/Chip.tsx) |
 | Card | `outlined` (default) · `brand` · `elevated` | — | — | — | [`Card.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/Card/Card.tsx) |
-| ServiceTag | `text` (default) · `icon-text` · `icon` | — | `sm` · `md` · `lg` | — | [`ServiceTag.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/ServiceBadge/ServiceTag.tsx) |
+| ServiceTag | `text` (default) · `icon-text` · `icon` | — | `sm` · `md` · `lg` | — | [`ServiceTag.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/ServiceTag/ServiceTag.tsx) |
 | Tag | — | `solid` (default) · `subtle` | `xs` · `sm` · `md` · `lg` | — | [`Tag.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/Tag/Tag.tsx) |
 | Badge | — | `solid` (default) · `subtle` | `xs` · `sm` · `md` · `lg` | `positive` · `warning` · `error` · `info` · `progress` · `brand` | [`Badge.tsx`](https://github.com/brikdesigns/brik-bds/blob/main/components/ui/Badge/Badge.tsx) |
